+++ orttraining/orttraining/core/graph/graph_parser/pattern_graph.cc
@@ -1,4 +1,4 @@
-#include "orttraining/core/graph/graph_parser.h"
+#include "orttraining/core/graph/graph_parser/pattern_graph.h"
 
 #include <queue>
 
@@ -145,86 +145,6 @@ bool PatternGraph::find_match(const Node* g, const Node* p, std::unordered_set<c
 
   PARSER_MSG(std::cout << "return true." << std::endl;)
   return true;
-}
-
-Status PatternGraph::TryMatch(Graph& target_graph, std::vector<PNN>& res,
-                              const std::unordered_set<std::string>* p_initializer_names_to_preserve) {
-  ORT_RETURN_IF_ERROR(to_graph(model, p_initializer_names_to_preserve));
-  auto& pattern_graph = model.MainGraph();
-
-  GraphViewer graph_viewer(target_graph);
-
-  GraphViewer pattern_viewer(pattern_graph);
-  const auto& graph_topology_list = graph_viewer.GetNodesInTopologicalOrder();
-  const auto& pattern_topology_list = pattern_viewer.GetNodesInTopologicalOrder();
-
-  Node* pattern_root = nullptr;  // Not real root, only a root specified by user
-  if (pattern_topology_list.size() && root_node.empty()) {
-    pattern_root = pattern_graph.GetNode(0);
-  }
-  for (auto node_index : pattern_topology_list) {
-    auto* node = pattern_graph.GetNode(node_index);
-    if (strcmp(node->Name().c_str(), root_node.c_str()) == 0) {
-      pattern_root = node;
-      break;
-    }
-  }
-  if (!pattern_root) {
-    return Status(common::ONNXRUNTIME, common::FAIL, "Pattern root was not found.");
-  }
-
-  for (auto node_index : graph_topology_list) {
-    auto* node = target_graph.GetNode(node_index);
-    res.clear();
-    std::unordered_set<const Node*> graph_path, pattern_path;
-    std::unordered_map<const Node*, const Node*> path_map;
-    if (find_match(node, pattern_root, graph_path, pattern_path, path_map, res, target_graph, pattern_graph)) {
-      res.push_back({node_index, pattern_root});
-      return Status::OK();
-    }
-  }
-
-  return Status(common::ONNXRUNTIME, common::FAIL, "No match for the target graph.");
-}
-
-Status PatternGraph::TryReplace(Graph& graph, const NodeDef& alternative, std::vector<std::pair<std::string, int>> fusion_inputs,
-                                std::vector<std::pair<std::string, int>> fusion_outputs,
-                                const std::unordered_set<std::string>* p_initializer_names_to_preserve) {
-  std::vector<PNN> match_results;
-  ORT_RETURN_IF_ERROR(TryMatch(graph, match_results, p_initializer_names_to_preserve));
-  InlinedVector<std::reference_wrapper<Node>> matched_nodes;
-  for (auto iter = match_results.rbegin(); iter != match_results.rend(); iter++) {
-    auto node = graph.GetNode(iter->first);
-    matched_nodes.push_back(*node);
-  }
-
-  auto add_node_args_with_name = [&match_results](Graph& graph, std::string name, int idx, std::vector<NodeArg*>& args) {
-    GraphViewer viewer(graph);
-    // const auto& node_topology_list = viewer.GetNodesInTopologicalOrder();
-    for (auto [node_index, pattern_node] : match_results) {
-      if (pattern_node->Name() == name) {
-        auto target_node = graph.GetNode(node_index);
-        args.push_back(target_node->MutableInputDefs()[idx]);
-        return Status::OK();
-      }
-    }
-    return Status(common::ONNXRUNTIME, common::FAIL);
-  };
-
-  std::vector<NodeArg*> input_args, output_args;
-  for (auto item : fusion_inputs) {
-    ORT_RETURN_IF_ERROR(add_node_args_with_name(graph, item.first, item.second, input_args));
-  }
-  for (auto item : fusion_outputs) {
-    ORT_RETURN_IF_ERROR(add_node_args_with_name(graph, item.first, item.second, output_args));
-  }
-  Node& replace_node = graph.AddNode(alternative.name, alternative.op_type, "",
-                                     input_args, output_args, {}, alternative.domain.c_str());
-  for (auto& attr : alternative.attributes) {
-    replace_node.AddAttribute(attr.first, attr.second);
-  }
-  graph_utils::FinalizeNodeFusion(graph, matched_nodes, replace_node);
-  return Status::OK();
 }
 
 }  // namespace training
